chore: add functionality to the buttons on the onboarding screen to redirect to the login screen
make the main screen the onboarding screen to avoid unnecessary redirect
